test functions swiched to router view
